+++ src/Core/Plugin.php
@@ -175,6 +175,21 @@ class Plugin
     }
 
     /**
+    * Gets a path for a plugin
+    *
+    * @param string $plugin
+    * @return string
+    * @throws \Origin\Core\Exception\MissingPluginException
+    */
+    public static function path(string $plugin) : string
+    {
+        if (! isset(static::$loaded[$plugin])) {
+            throw new MissingPluginException($plugin);
+        }
+        return static::$loaded[$plugin]['path'];
+    }
+
+    /**
      * Includes a file
      *
      * @param string $filename